AD provider: add Smart Card handling
This adds smart card handling to the ad-open-provider module.

When requested by the client, the AD provider will attempt to
load and initialize Cryptoki library, scan all slots with active
tokens in the system, fetch the certificates and retrieve UPNs
from the certificates.

The UPNs are then used to see if users exist on the domain and if
one and only one valid user is found, the data is returned to the
client so that it can ask for PIN and continue authentication
(users are expected to have only one card inserted at a time).

QA Notes:
Testing Done:
Built likewise and tried various cards with coolkey library as Cryptoki
implementation, sandbox of likewise.
Documentation Notes:
Bug Number:
Reviewed by: arolett, krishnag, snambakam
Approved by:
Mailto:

[git-p4: depot-paths = "//public/vmware-likewise-open-6-1/main/": change = 24183]