add likewise reg database to build
Currently on ESX we need to import the registry setting via
the command "lwregshell import file". This needs to happen
after lwsmd is running and the lwreg process within it is up.
We also need to do a "lwsmd reload" afterwards.

This change simply puts the imported version of the database
for the embedded build on the system. This is to speed up
testing and development. If there are changes to the reg
entries, then this file needs to be re-generated and
re-checked in. Something that will affect this is if we modify
the path for embedded build. These path are encoded in the reg
files.

We may need to eventually find a better way to do this. Either
import the setting on ESX first boot, or code it in the
vmwauth wrapper library to do the import once.

QA Notes:
Testing Done: tested on ESX.
Documentation Notes:
Bug Number:
Reviewed by:
Approved by:
Mailto:

[git-p4: depot-paths = "//public/vmware-likewise-open-6-1/main/": change = 24503]