Smart Card: fix PAM to properly handle non-smart card case
When smart card support is enabled in pam_lsass, but user
who is trying to log in is not the one that certificate on
the smart card is issued to PAM layer would get confused
and would accept any password.

Split pam_sm_authenticate() into manageable chunks so that
logic flow is clear and ensured that we drop into password
based authentication path if there is no card in the reader
or if service calling (login, ssh, etc) is not on the list
of services that should be using smart card authentication.

QA Notes:
Testing Done:
Copied pam_libsass.so over and verified that I can't log in
with blank/garbage password anymore, sandbox of likewise ESXi
build.
Documentation Notes:
Bug Number:
Reviewed by: arolett, grantr
Approved by:
Mailto:

[git-p4: depot-paths = "//public/vmware-likewise-open-6-1/main/": change = 31430]