Preventing memory leak by not using list head as iterator.
A memory leak was caused by using the head of a linked list as an iterator. The list's head was not restored when errors ocurred in addition to somewhere in the code where a previous developer forgot to restore it.

QA Notes:
Testing Done:
Documentation Notes:
Bug Number: 1145367
Reviewed by: snambakam
Approved by: krishnag
Mailto: vmidentity-checkins

[git-p4: depot-paths = "//public/vmware-likewise-open-6-1/main/": change = 38247]