Match smart card user against canonicalized user name.
The user object returned by LsaGetSmartCardUserObject contains the
canonicalized user name for the user matching the UPN from the
certificate found on the card. To make the comparison to the user
being authenticated valid the canonicalized version of the name
provided to PAM is required for comparison.

This fixes the case were a user attempts to log in as user@domain
or user@DOMAIN as the canonicalized name is DOMAIN\user.

QA Notes:
Testing Done: Attempted to log in as user@domain, user@fqdn, user@DOMAIN, user@FQDN, DOMAIN\user and FQDN\user where domain is the short form of the AD domain (pib-scr) and fqdn is the fully qualified AD domain (pib-scr.eng.vmware.com) and capitalization indicates an all upper-case or lower-case domain name. All were successful. Attempting to log in with the wrong domain or an incorrect user resulted in an password prompt (instead of PIN) as expected. When the user was valid but did not match the card the password prompt would only appear after a delay caused by the requests to the card for verification.
Documentation Notes:
Bug Number:
Reviewed by: dtor ssalley
Approved by:
Mailto:

[git-p4: depot-paths = "//public/vmware-likewise-open-6-1/main/": change = 40264]