Remove <PRINCIPAL> matching requirement from PKINIT configuration.
The <PRINCIPAL> matching requirement appears to be a feature added to
Kerberos during Likewise development. It is only applicable to smart
card authentication but because smart card support was not released
with Likewise Open there is no way to know how it was intended to
work. Requiring the pricipal name to match the UPN on the certificate
used for authentication prevents DoD Common Access Cards from working
because the principal name, which depends on the domain credential
being acquired, does not always match the UPN on the card. The UPN
may be listed as an alternative name on the account, a concept
supported by Active Directory but not by Kerberos.

No third party instructions for enabling CAC or other smart cards
with Kerberos include mention of this extra matching criteria, only
the "<EKU>msScLogin" matching rule is usually given, sometimes with
an extra requirement that the certificate support the digitalSign
key usage as well.

The result of removing this matching criteria will be that a card
with multiple valid login certificates may not function because pkinit
will select the wrong certificate. As mentioned above, there is no
evidence for such cards being in wide circulation.

QA Notes:
Testing Done: Authenticated to Active Directory using a CAC.
Documentation Notes:
Bug Number:
Reviewed by: dtor
Approved by:
Mailto:

[git-p4: depot-paths = "//public/vmware-likewise-open-6-1/main/": change = 40266]