Likewise: can't login as root when joined to domain
There's a small bug in likewise where, if you're joined to a domain and
a smartcard is inserted, logging in as root no longer works. It tries
to lookup root as a domain user, and instead of failing it succeeds,
but when it asks for your password it obviously can't match it against
anything, so the login is rejected. We should return "not handled" when
root cannot be found in the domain, so that a local lookup is
attempted.

QA Notes:
Testing Done:
o Joined ESX box to PIB-SCR domain, inserted smartcard and logged in
  as root and also as a domain user
o https://sandbox.eng.vmware.com/sb/3747756/ (vmware-likewise-open-esx-x32)
Documentation Notes:
Bug Number: 1275903
Reviewed by: arolett
Approved by:
Mailto:

[git-p4: depot-paths = "//public/vmware-likewise-open-6-1/main/": change = 42894]