Modify vmdir provider to properly interpret new SID format
Modifying the SID format, adding a new MachineID field, broke some core
assumptions made by the vmdir provider. The most basic: the "RID" field
is safe to use as the UNIX uid value. Previously, the RID field was
the (MachID | RID) represented as a DWORD. The SID format change split
out the MachID as a separate "sub authority" field. This is bad, as the
RID value now is a low number (1024 -> N) which overlaps with UNIX
/etc/password assigned values.

This change modifies VmDirGetRID() to recognize the new SID format, and
the returned "RID" value is 8-bits of the MachineID or'ed in with the
RID field. Result: same "RID" value returned before the SID format change.
The new function VmDirGetRIDFromUID() performs the inverse mapping. Given
a UID value, which is a value returned from VMDirGetRID, it returns a
string of the format "M-RID", or just "RID" if the high byte is not set.

This solution restores full functionality of the vmdir provider with the
new SID format change in lotus/main. This solution is a compromise, as it
limits the total number of users to 16 million, and the maximum allowed
MachID space to 255 entries. A better UID->User SID mapping solution must
be designed to solve the general problem. This is an interum solution to
restore vmdir provider functionality.

Trivial changes to provider-main.c to populate "Domain/DomainSID" information.

QA Notes:
Testing Done: local build, add identity to vmdir, configure nsswitch to use
	lsass, run "id newUser". Also am able to ssh DOMAIN\\user and su DOMAIN\\user.
	SB Build: https://buildweb.eng.vmware.com/sb/5598732/
Documentation Notes:
Bug Number: 1457395
Reviewed by: ssalley
Approved by: krishnag
Mailto: brownj, ssalley, snambakam

[git-p4: depot-paths = "//public/vmware-likewise-open-6-1/main/likewise-v6.1/": change = 63664]