Handle SDDL strings with no ACEs in DACL.
Note that there's a bug in RtlpParseSddlString that causes the last
character of the DACL to be dropped. This is fine if there's an ACE
as we lose the closing paren and the other code seems to handle that
fine. However, in the case of "AI" with no subsequent ACE we get back
just "A" and bail (as that's not a valid flag). That parsing bug will
be dealt with separately (and is a bigger, more dangerous change). This
is just to ensure that we don't crash.

Change-Id: I693a8029cee3945e420b65b8c1deeac65c798007