Improve scenario of Domains going OFFLINE intermittently due to network loss resulting in authentication failures
We often see the domains going OFFLINE in customer environments which are intermittent due to network loss, however,
the OFFLINE state is cached in lsass daemon and all future requests attempts to authenticate with the domain are
failed. The fix is to re-attempt connection with the domain when we see an OFFLINE state. The lsass ad-provider will
attempt to transition ONLINE, if it can. If not, it will fail with LW_ERROR_DOMAIN_IS_OFFLINE which is a more permanent failure.

Change-Id: Ic706e68fb0759e48fe9cb873a18b402b23608342