Add memory cache to vmdir provider
This change adds a memory caching layer to the vmdir provider for
user, groups, and group memberships.  After an object is queried
from LDAP it is stored in the cache.  Subsequent queries for the
object may then be resolved from the cache (if not expired).  This
improves the performance of the vmdir provider operations, in
particular for creating access tokens.

A new registry key has been added to allow configuration of the
expiration time in seconds:

[Services\lsass\Parameters\Providers\VmDir\]
   "CacheEntryExpiry" REG_DWORD       0x00003840 (14400)

The default of 14400 seconds is 4 hours.

PR#1926648

Change-Id: I704a2ef06c9dca0481606db12a91502a5e4b813a