Binary SID changes
Adjustments to accomodate a Binary SID return type vs a string. SIDs in
Active Directory are stored in Binary format, not the string representation.
Vmdird ldap-head/result.c is modified to return a binary SID when compiled
with --enable-winjoin. These changes to likewise-open are needed to
handle either a string result or a binary result. This implementation choice
avoids the need for a build switch like --enable-winjoin.

Change-Id: I6161b8deb6dd1df3b11edea485122076ed81e446