Fix security descriptor string conversion
This changes fixes a problem with the conversion of SIDs
to SDDL string format.  It was incorrectly converting some
SID strings to builtin abbreviations such as BU.  The
conversion is now done by looking up the SDDL table by
SID string instead of by RID value.

Also modified a log message in the lsass vmdir provider to
not log an unset value, and updated RPM package release.

Change-Id: Ic829adf21a58dd53ee57f402234323a1d98542ca