Open firewall port for drsuapi RPC service
Call helper script (firewall-drsuapi.sh) to open
the ephererial port which the drsuapi RPC service
was assigned. This service is discovered by the DCE/RPC
endpoint mapper. As this is not a fixed port, an incoming
"hole" throught the filewall must be opened. My research
indicates there is no iptables API to call to accomplish this
task, so calling system() to a helper script appears to be
the only option.

Change-Id: I92deddac5dcae052e132eae3736548ce95d95667