Fix likewise-open RPM upgrade
Upgrade install of the likewise-open RPM failed
because all the likewise services were still running
when the new binaries were copied to disk.

This change modifies the pre-install scriptlet to
stop lwsmd and all the services it manages when invoked
with the upgrade argument.  The post-install scriptlet
for upgrade now performs "systemctl daemon-reload".

The systemctl configuration for lwsmd now uses type
"forking".

Update RPM patch level to 10.

Change-Id: I9750a49e683ead3ced657687a2881fd83edff8fe